requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.4.0
pyotp==2.8.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
